# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ARC — Rules & Calculations
  version: 0.2.0
  description: Unauthenticated authoring and execution API. Published graphs and their pinned dependencies are immutable.
servers:
  - url: http://localhost:8080/api
  - url: http://localhost:3080/api
paths:
  /rule-summaries:
    parameters:
      - {$ref: '#/components/parameters/Offset'}
      - {$ref: '#/components/parameters/Limit'}
      - {name: search, in: query, schema: {type: string, maxLength: 200}}
      - {name: kind, in: query, schema: {type: string, enum: [DECISION_TREE, FORMULA, RULE]}}
      - {name: publishedOnly, in: query, schema: {type: boolean, default: false}}
    get:
      operationId: listRuleSummaries
      responses:
        '200':
          description: Bounded page of metadata without graph or source definitions
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/PageMetadata'}
                  - type: object
                    properties:
                      items:
                        type: array
                        items: {$ref: '#/components/schemas/RuleSummary'}
        default: {$ref: '#/components/responses/Error'}
  /source-summaries:
    parameters:
      - {$ref: '#/components/parameters/Offset'}
      - {$ref: '#/components/parameters/Limit'}
      - {name: search, in: query, schema: {type: string, maxLength: 200}}
    get:
      operationId: listSourceSummaries
      responses:
        '200':
          description: Bounded page of metadata without graph or source definitions
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/PageMetadata'}
                  - type: object
                    properties:
                      items:
                        type: array
                        items: {$ref: '#/components/schemas/SourceSummary'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}/version-summaries:
    parameters:
      - {$ref: '#/components/parameters/RuleId'}
      - {$ref: '#/components/parameters/Offset'}
      - {$ref: '#/components/parameters/Limit'}
    get:
      operationId: listRuleVersionSummaries
      responses:
        '200':
          description: Bounded page of metadata without graph or source definitions
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/PageMetadata'}
                  - type: object
                    properties:
                      items:
                        type: array
                        items: {$ref: '#/components/schemas/VersionSummary'}
        default: {$ref: '#/components/responses/Error'}
  /sources/{id}/version-summaries:
    parameters:
      - {$ref: '#/components/parameters/RuleId'}
      - {$ref: '#/components/parameters/Offset'}
      - {$ref: '#/components/parameters/Limit'}
    get:
      operationId: listSourceVersionSummaries
      responses:
        '200':
          description: Bounded page of metadata without graph or source definitions
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/PageMetadata'}
                  - type: object
                    properties:
                      items:
                        type: array
                        items: {$ref: '#/components/schemas/SourceVersionSummary'}
        default: {$ref: '#/components/responses/Error'}
  /rules:
    get:
      operationId: listRules
      responses:
        '200':
          description: Rules with current drafts
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Rule'}
    post:
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, kind]
              properties:
                id: {type: string, pattern: '^[a-z][a-z0-9-]{0,79}$'}
                name: {type: string, minLength: 1, maxLength: 160}
                description: {type: string, maxLength: 2000}
                kind: {type: string, enum: [DECISION_TREE, FORMULA, RULE]}
                definition: {$ref: '#/components/schemas/Definition'}
      responses:
        '201':
          description: Created rule (definition defaults to a template)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}:
    parameters: [{$ref: '#/components/parameters/RuleId'}]
    get:
      operationId: getRule
      responses:
        '200':
          description: Rule and draft
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: saveDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, revision, definition]
              properties:
                name: {type: string, minLength: 1, maxLength: 160}
                description: {type: string, maxLength: 2000}
                revision: {type: integer, minimum: 1}
                definition: {$ref: '#/components/schemas/Definition'}
      responses:
        '200':
          description: Saved rule with incremented revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}/publish:
    parameters: [{$ref: '#/components/parameters/RuleId'}]
    post:
      operationId: publishRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision: {type: integer, minimum: 1}
      responses:
        '200':
          description: Rule with incremented revision and publishedVersion
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rule'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}/execute:
    parameters: [{$ref: '#/components/parameters/RuleId'}]
    post:
      operationId: executeRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputs]
              properties:
                inputs: {type: object, additionalProperties: true}
                version: {type: integer, minimum: 1, description: Omit for latest published version}
                trace: {type: boolean, nullable: true, default: true, description: Include a bounded execution trace; null uses the default}
                timeoutMs: {type: integer, nullable: true, minimum: 100, maximum: 30000, default: 30000, description: Shared execution deadline in milliseconds; null uses the default}
            example:
              inputs: {orderTotal: 150, customerTier: premium}
              version: 1
      responses:
        '200': {$ref: '#/components/responses/Execution'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}/versions:
    parameters: [{$ref: '#/components/parameters/RuleId'}]
    get:
      operationId: listVersions
      responses:
        '200':
          description: Versions, newest first
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Version'}
        default: {$ref: '#/components/responses/Error'}
  /rules/{id}/versions/{version}:
    parameters:
      - {$ref: '#/components/parameters/RuleId'}
      - name: version
        in: path
        required: true
        schema: {type: integer, minimum: 1}
    get:
      operationId: getVersion
      responses:
        '200':
          description: Immutable graph snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Version'}
        default: {$ref: '#/components/responses/Error'}
  /variables:
    post:
      operationId: availableVariables
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Definition'}
      responses:
        '200':
          description: Guaranteed available variables by node ID, including incomplete acyclic drafts
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items: {type: string}
        default: {$ref: '#/components/responses/Error'}
  /validate:
    post:
      operationId: validateGraph
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Definition'}
      responses:
        '200':
          description: Graph is structurally valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: {type: boolean, enum: [true]}
        default: {$ref: '#/components/responses/Error'}
  /preview:
    post:
      operationId: previewGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [definition, inputs]
              properties:
                trace: {type: boolean, nullable: true, default: true, description: Include a bounded execution trace; null uses the default}
                timeoutMs: {type: integer, nullable: true, minimum: 100, maximum: 30000, default: 30000, description: Shared execution deadline in milliseconds; null uses the default}
                definition: {$ref: '#/components/schemas/Definition'}
                inputs: {type: object, additionalProperties: true}
      responses:
        '200': {$ref: '#/components/responses/Execution'}
        default: {$ref: '#/components/responses/Error'}
  /functions:
    get:
      operationId: listFunctions
      responses:
        '200':
          description: Runtime function catalog, including explicit supported flags and insertion snippets
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/Function'}}
  /studio/expression/check:
    post:
      operationId: checkExpression
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expression]
              properties:
                expression: {type: string, maxLength: 2000}
      responses:
        '200':
          description: Syntax/dependency check without execution or external data access
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: {type: boolean}
                  variables: {type: array, items: {type: string}}
                  error: {type: string, nullable: true}
  /studio/build:
    post:
      operationId: buildArcScript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: {type: string, maxLength: 100000}
      responses:
        '200':
          description: Compiled definition or diagnostics with one-based line and column; syntax errors have a null definition
          content:
            application/json:
              schema:
                type: object
                properties:
                  definition: {$ref: '#/components/schemas/Definition'}
                  source: {type: string}
                  diagnostics:
                    type: array
                    items:
                      type: object
                      properties:
                        message: {type: string}
                        line: {type: integer}
                        column: {type: integer}
  /studio/render:
    post:
      operationId: renderArcScript
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Definition'}
      responses:
        '200':
          description: Canonical ARC Script preserving graph identity, layout, bindings, and comments
          content:
            application/json:
              schema: {type: object, properties: {source: {type: string}}}
  /sources:
    get:
      operationId: listSources
      responses:
        '200':
          description: Latest data source versions
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DataSource'}}
    post:
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, definition]
              properties:
                id: {type: string, pattern: '^[a-z][a-z0-9-]{0,79}$'}
                name: {type: string, maxLength: 160}
                definition: {$ref: '#/components/schemas/SourceConfig'}
      responses:
        '200':
          description: Source created with configuration version 1
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
  /sources/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    put:
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, revision, definition]
              properties:
                name: {type: string, maxLength: 160}
                revision: {type: integer, description: Current source version for optimistic concurrency}
                definition: {$ref: '#/components/schemas/SourceConfig'}
      responses:
        '200':
          description: New immutable configuration version
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
        '409': {description: Stale revision}
  /sources/{id}/versions:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    get:
      operationId: sourceVersions
      responses:
        '200':
          description: Configuration history, newest first
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/DataSource'}}
  /sources/{id}/versions/{version}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
      - {name: version, in: path, required: true, schema: {type: integer}}
    get:
      operationId: sourceVersion
      responses:
        '200':
          description: Pinned source configuration
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DataSource'}
  /sources/{id}/test:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    post:
      operationId: testSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputs]
              properties:
                inputs: {type: object, additionalProperties: true}
                version: {type: integer, minimum: 1}
      responses:
        '200':
          description: Source response, before JSON Pointer extraction
          content:
            application/json:
              schema: {type: object, properties: {result: {nullable: true}}}
components:
  parameters:
    Offset:
      name: offset
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    RuleId:
      name: id
      in: path
      required: true
      schema: {type: string}
  responses:
    Error:
      description: 400 malformed request; 404 missing resource; 409 revision conflict, duplicate ID, or unpublished rule; 413 oversized request; 422 invalid graph/input/calculation; 500 server error; 504 execution deadline exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              status: {type: integer}
              message: {type: string}
              issues: {type: array, items: {type: string}}
              locations:
                type: array
                items:
                  type: object
                  properties:
                    ruleId: {type: string, nullable: true}
                    version: {type: integer, nullable: true}
                    nodeId: {type: string}
                    label: {type: string}
    Execution:
      description: Complete result, optional bounded trace, and preparation/execution timing
      content:
        application/json:
          schema:
            type: object
            properties:
              ruleId: {type: string}
              version: {type: integer, nullable: true}
              result: {nullable: true, description: A single reached Output value or an object keyed by node ID when multiple Outputs are reached}
              durationMicros: {type: integer, format: int64, description: Engine duration excluding preparation and response serialization}
              traceEnabled: {type: boolean}
              traceTruncated: {type: boolean, description: True when the 256 KiB trace budget stopped recording complete steps}
              executedSteps: {type: integer, minimum: 0, maximum: 1000}
              traceBytes: {type: integer, minimum: 2, maximum: 262144, description: UTF-8 JSON bytes of the retained trace array}
              timing:
                type: object
                properties:
                  preparationMicros: {type: integer, format: int64}
                  executionMicros: {type: integer, format: int64}
                  totalMicros: {type: integer, format: int64, description: Server preparation and execution, excluding HTTP serialization and transfer}
              sources:
                type: array
                items:
                  type: object
                  properties:
                    input: {type: string}
                    sourceId: {type: string}
                    version: {type: integer}
                    status: {type: string, enum: [RESOLVED, DEFAULT]}
                    durationMicros: {type: integer, format: int64}
              trace:
                type: array
                items:
                  type: object
                  properties:
                    ruleId: {type: string}
                    version: {type: integer, nullable: true}
                    nodeId: {type: string}
                    label: {type: string}
                    type: {type: string}
                    value: {nullable: true}
                    branch: {type: string, nullable: true, enum: [next, 'true', 'false', null]}
                    depth: {type: integer}
  schemas:
    PageMetadata:
      type: object
      required: [items, total, offset, limit]
      properties:
        total: {type: integer, format: int64, minimum: 0}
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
    RuleSummary:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        kind: {type: string, enum: [DECISION_TREE, FORMULA, RULE]}
        revision: {type: integer}
        publishedVersion: {type: integer, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        nodeCount: {type: integer}
        inputCount: {type: integer}
        referenceCount: {type: integer}
    VersionSummary:
      type: object
      properties:
        ruleId: {type: string}
        version: {type: integer}
        publishedAt: {type: string, format: date-time}
    SourceSummary:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        version: {type: integer}
        kind: {type: string, enum: [HTTP, LOOKUP]}
    SourceVersionSummary:
      type: object
      properties:
        id: {type: string}
        version: {type: integer}
        createdAt: {type: string, format: date-time}
    Rule:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        description: {type: string}
        kind: {type: string, enum: [DECISION_TREE, FORMULA, RULE]}
        draft: {$ref: '#/components/schemas/Definition'}
        revision: {type: integer}
        publishedVersion: {type: integer, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    Version:
      type: object
      properties:
        ruleId: {type: string}
        version: {type: integer}
        definition: {$ref: '#/components/schemas/Definition'}
        publishedAt: {type: string, format: date-time}
    Definition:
      type: object
      required: [schemaVersion, inputs, nodes, edges]
      properties:
        notes: {type: array, nullable: true, maxItems: 500, items: {type: string, maxLength: 2000}}
        schemaVersion: {type: integer, enum: [1]}
        inputs:
          type: array
          maxItems: 50
          items:
            type: object
            required: [name, type, required]
            properties:
              name: {type: string, pattern: '^[A-Za-z_][A-Za-z_0-9]{0,63}$'}
              type: {type: string, enum: [NUMBER, STRING, BOOLEAN, ARRAY, OBJECT]}
              required: {type: boolean}
              defaultValue: {nullable: true}
              source: {$ref: '#/components/schemas/SourceBinding'}
        nodes:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [id, type, label]
            properties:
              id: {type: string, pattern: '^[A-Za-z0-9_-]{1,80}$'}
              type: {type: string, enum: [INPUT, FORMULA, CONDITION, SWITCH, TRANSFORM, REFERENCE, OUTPUT]}
              label: {type: string, minLength: 1, maxLength: 160}
              position:
                type: object
                properties:
                  x: {type: number}
                  y: {type: number}
              expression: {type: string, maxLength: 2000, nullable: true}
              output: {type: string, nullable: true}
              ruleId: {type: string, nullable: true}
              version: {type: integer, minimum: 1, nullable: true}
              cases:
                type: array
                nullable: true
                maxItems: 20
                description: SWITCH only; ordered first-match boolean cases, with stable case-sensitive IDs
                items:
                  type: object
                  required: [id, label, expression]
                  properties:
                    id: {type: string, pattern: '^[A-Za-z0-9_-]{1,64}$'}
                    label: {type: string, minLength: 1, maxLength: 160}
                    expression: {type: string, maxLength: 2000}
              fields:
                type: array
                nullable: true
                maxItems: 50
                description: TRANSFORM only; unique literal field names; cannot combine nonempty fields with expression
                items:
                  type: object
                  required: [name, expression]
                  properties:
                    name: {type: string, minLength: 1, maxLength: 160}
                    expression: {type: string, maxLength: 2000}
              bindings:
                type: object
                nullable: true
                maxProperties: 50
                additionalProperties: {type: string, maxLength: 2000}
        edges:
          type: array
          maxItems: 200
          items:
            type: object
            required: [id, source, target, sourceHandle]
            properties:
              id: {type: string, maxLength: 100}
              source: {type: string}
              target: {type: string}
              sourceHandle: {type: string, pattern: '^(next|true|false|default|case:[A-Za-z0-9_-]{1,64})$'}

    SourceBinding:
      type: object
      nullable: true
      required: [id, version, bindings, onError]
      properties:
        id: {type: string}
        version: {type: integer, minimum: 1}
        bindings: {type: object, maxProperties: 20, additionalProperties: {type: string, maxLength: 2000}}
        pointer: {type: string, maxLength: 500, description: JSON Pointer; empty selects the entire value}
        onError: {type: string, enum: [FAIL, DEFAULT]}
    SourceConfig:
      type: object
      required: [kind, parameters, timeoutMs]
      properties:
        kind: {type: string, enum: [LOOKUP, HTTP]}
        url: {type: string, maxLength: 2000}
        parameters:
          type: array
          maxItems: 20
          items:
            type: object
            required: [name, type, required]
            properties:
              name: {type: string}
              type: {type: string, enum: [NUMBER, STRING, BOOLEAN]}
              required: {type: boolean}
              defaultValue: {nullable: true}
        entries: {type: object, maxProperties: 1000, additionalProperties: true}
        secretHeaders: {type: object, maxProperties: 10, additionalProperties: {type: string}, description: Header names mapped to ARC_SECRET_ environment aliases; never raw credentials}
        timeoutMs: {type: integer, minimum: 100, maximum: 10000}
    DataSource:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        version: {type: integer}
        definition: {$ref: '#/components/schemas/SourceConfig'}
    Function:
      type: object
      properties:
        name: {type: string}
        category: {type: string}
        signature: {type: string}
        description: {type: string}
        snippet: {type: string}
        supported: {type: boolean}
        origin: {type: string}